renamed addons to packages
